extract client.utils package
