+++ databend-client/src/main/java/com/databend/client/utils/JsonCodec.java
@@ -12,8 +12,9 @@
  * limitations under the License.
  */
 
-package com.databend.client;
+package com.databend.client.utils;
 
+import com.databend.client.DiscoveryResponseCodec;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
