+++ databend-jdbc/src/main/java/com/databend/jdbc/DatabendDriverUri.java
@@ -24,7 +24,7 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
-import static com.databend.client.OkHttpUtils.*;
+import static com.databend.client.utils.OkHttpUtils.*;
 import static com.databend.jdbc.ConnectionProperties.*;
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Strings.isNullOrEmpty;
